Fixed bug with multiple fxml files on one controller class.
