+++ src/main/resources/ConnectWindow.fxml
@@ -6,7 +6,7 @@
 <?import javafx.scene.layout.AnchorPane?>
 <?import javafx.scene.layout.Pane?>
 
-<AnchorPane xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.benjazzy.mixchat.MixController">
+<AnchorPane xmlns="http://javafx.com/javafx/8.0.171" xmlns:fx="http://javafx.com/fxml/1" fx:controller="com.benjazzy.mixchat.controller.ConnectController">
    <children>
       <Pane prefHeight="98.0" prefWidth="200.0">
          <children>
